Dismiss RN view when 'exit' button is pressed
It seems like F3HViewController's 'self' isn't accessible in the native
method we're exposing to RN. I.e. This doesn't work when the method is
called by RN:

  [self dismissViewControllerAnimated:YES completion:nil]

However, we _can_ access the AppDelegate from the exposed native method.
So we access the RN root view by making a property on the AppDelegate.

Hat Tip: https://moduscreate.com/blog/leverage-existing-ios-views-react-native-app/